refactor(viewer): extract georef-align module + thread LoadTarget (no behavior change)
Scaffolding for the canonical single load path (no behavior change):

- Extract the georeferencing/alignment cluster (ModelGeoref + extractModelGeoref
  + alignGeometryToReference + findReferenceGeorefModel + private transform
  helpers) out of useIfcFederation.ts into a shared ingest/federationAlign.ts,
  so the upcoming finalizeModel in useIfcLoader can call it without a circular
  import. Also de-dups realignFederation. Pure move; useIfcFederation imports
  the three entry points back.

- Give useIfcLoader.loadFile an optional `target: LoadTarget` (default
  `{kind:'primary'}`). Source the model id from the target, and gate
  resetViewerState()/clearAllModels() on `target.kind==='primary'` so a future
  federated add can never wipe model #1. Rename the in-scope primaryModelId →
  modelId. With the default target this path is byte-identical to before.

Typecheck clean.
